v0.1.0: CXXStateTree FSM, dual-screen pipeline, replay integration test (#2)
* Initial implementation

* Add markdown about transition state

* Simplifying transitions

* Remove wildcard

* Removed watchdog

* Add screen mode + per-screen detection types

* Enforce strict unified YAML schema for top/bottom blocks

* Implement single/dual screen evaluation semantics

* Plumb top+bottom ROI maps through preprocessing and runtime

* Migrate xy starter hunt to new schema

* Update for strict schema, dual-screen logic, and pipeline plumbing

* Formatting fixes

* Add color improvement and sudden intensity drop detector

* Add color correction to the pipeline

* Add XY Starter Fennekin replay integration test and dual-screen FSM

* Remove console mode from entry point, GUI only

* Fix intensity detector advanced multiple times per frame

* Remove trailing underscores from IntensityEventDetector members

* Cache CLAHE and gamma LUT in ImproveFrameColors, fix doc comment

* Eliminate double ROI extraction per frame in dual-screen pipeline

* Extract ParseInputActions helper to remove duplicate YAML parsing block

* Add code review command

* Update CHANGELOG and CLAUDE.md for v0.1.0

* Add ToHuntConfig unit test

* Fix GitHub Actions workflow permissions for Claude to post PR comments

+++ tests/unit/TestScreenDetector.cpp
@@ -174,7 +174,7 @@ TEST(ScreenDetector, TemporalSmoothingReducesJitter)
     auto &prev = detectedCorners[detectedCorners.size() - 2];
 
     float maxDelta = 0.0f;
-    for (int j = 0; j < 4; ++j)
+    for (size_t j = 0; j < 4; ++j)
     {
         maxDelta = std::max(maxDelta, std::abs(last[j].x - prev[j].x));
         maxDelta = std::max(maxDelta, std::abs(last[j].y - prev[j].y));
@@ -288,11 +288,7 @@ TEST(ScreenDetector, DetectsTrapezoidal5To3Screen)
     auto frame = MakeBlackFrame();
     // Perspective-distorted trapezoid with ~5:3 aspect ratio
     // Top edge narrower than bottom (camera below screen)
-    DrawBrightRect(frame,
-        cv::Point(250, 120),
-        cv::Point(650, 120),
-        cv::Point(700, 400),
-        cv::Point(200, 400));
+    DrawBrightRect(frame, cv::Point(250, 120), cv::Point(650, 120), cv::Point(700, 400), cv::Point(200, 400));
 
     SH3DS::Capture::ScreenDetector detector;
     auto result = detector.DetectOnce(frame);
